broken autocomplete

+++ Web/src/main/java/clientConverter.java
@@ -7,6 +7,7 @@ import javax.faces.convert.Converter;
 import javax.faces.convert.FacesConverter;
 import javax.inject.Inject;
 import javax.inject.Named;
+import java.math.BigDecimal;
 
 @Named
 @FacesConverter(value = "clientConverter", managed = true)
@@ -20,7 +21,7 @@ public class clientConverter implements Converter<Client> {
         Client[] clients= manageContractBean.getClients();
         int id = Integer.parseInt(value);
         for (Client client : clients) {
-            if (client.getDBB_ID().equals(id)) {
+            if (client.getDBB_ID().equals(BigDecimal.valueOf(id))) {
                 return client;
             }
         }
